View ids are no longer stored in the model
This finalized the view refactoring. These changes don't reduce code
as much as making it more "right" and easier extandible.
The view code used to be in the conquest state but most of it was
boilerplate repitition, now the boilerplate has its own file and the
view id's are basedd on the model in a seperate structure. Which makes
more sense if you think about it. For example the AI would've had
access to view knowledge, or worse, if ever mulitplayer was implemented
the view id sharing would be just weird.

This also allows recycling of view_id's. which prevents bloating of the
graph.